oslo-generator-shacl-template: ClassShapeBaseHander: add inheritance
SHACL shapes of each class are now targetting the class itself
and all its children. This way, inherited constraints on attributes of
the class are also enforced on all its children.

NOTE: redefined and subsetted attributes remain ignored. Those are
incorrectly inherited at the moment because they change the cardinality,
URI, or datatype of the attribute of the class in one of its children.
Support will be added in the future commits.